+++ src/main/java/com/thinkxfactor/springdemo/entityControllers/StudentEntityController.java
@@ -1,7 +1,7 @@
-package com.thinkxfactor.springdemo.controllers;
+package com.thinkxfactor.springdemo.entityControllers;
 
-import com.thinkxfactor.springdemo.entities.Book;
-import com.thinkxfactor.springdemo.repository.BookRepository;
+import com.thinkxfactor.springdemo.entities.Student;
+import com.thinkxfactor.springdemo.repository.StudentRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,31 +15,31 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/book")
+@RequestMapping("/student")
 @CrossOrigin
-public class BookController {
+public class StudentEntityController {
 
     @Autowired
-    BookRepository bookRepository;
+    StudentRepository studentRepository;
 
     // create
     @PostMapping("/create")
-    public ResponseEntity<?> create(@RequestBody Book book) {
-        return ResponseEntity.ok(bookRepository.save(book));
+    public ResponseEntity<?> create(@RequestBody Student student) {
+        return ResponseEntity.ok(studentRepository.save(student));
     }
 
     // read
     @GetMapping("/readAll")
     public ResponseEntity<?> readAll() {
-        return ResponseEntity.ok(bookRepository.findAll());
+        return ResponseEntity.ok(studentRepository.findAll());
     }
     @GetMapping("/readById")
     public ResponseEntity<?> readById(@RequestParam Long id) {
-        if(!bookRepository.existsById(id)) {
+        if(!studentRepository.existsById(id)) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
         else {
-            return ResponseEntity.ok(bookRepository.findById(id));
+            return ResponseEntity.ok(studentRepository.findById(id));
         }
     }
 }
