+++ src/main/java/com/thinkxfactor/springdemo/testControllers/TestController.java
@@ -1,4 +1,4 @@
-package com.thinkxfactor.springdemo.controllers;
+package com.thinkxfactor.springdemo.testControllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
